+++ plugins/core/qRANSAC_SD/RANSAC_SD_orig/LevMarFitting.h
@@ -122,8 +122,8 @@ bool LevMar(IteratorT begin, IteratorT end, FuncT &func,
 	typedef typename FuncT::ScalarType ScalarType;
 	enum { paramDim = FuncT::NumParams };
 	bool retVal = true;
-	unsigned int totalSize = end - begin;
-	if(!totalSize)
+	size_t totalSize = end - begin;
+	if (!totalSize)
 		return false;
 	ScalarType lambda = ScalarType(0.0001);
 	ScalarType *F0 = new ScalarType[totalSize * paramDim];
@@ -148,7 +148,7 @@ bool LevMar(IteratorT begin, IteratorT end, FuncT &func,
 	for(size_t i = subsetSizes.size(); i;)
 	{
 		--i;
-		subsetSizes[i] = totalSize;
+		subsetSizes[i] = static_cast<unsigned>(totalSize);
 		if(i)
 			subsetSizes[i] = subsetSizes[i] >> 1;
 		totalSize -= subsetSizes[i];
